% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuous_wavelet_transform.R
\name{cwt}
\alias{cwt}
\title{Apply the continuous wavelet transform to a signal}
\usage{
cwt(obs, scales, psi = "mexh")
}
\arguments{
\item{obs}{vector of the time dependent signal.  In our case, intensities.}

\item{scales}{vector of the wavelet scales which to apply.}

\item{psi}{the wavelet to apply. "mexh", "nmexh" or a data.frame containing the wavelet.}
}
\value{
A matrix, rows corresponding to wavelet scales, columns to time.
}
\description{
\code{cwt} returns a matrix of wavelet coefficients for the specified scales.
}
\details{
This function is a modification of the original cwt code provided by the centWave algorithm [1], [2] and [3].
}
\section{Attribution}{

This code was modified from the originally published centWave algorithm [1].  The code was orginally distributed and obtained under the GPL2 license via the xcms software package [2]. The original algorithms depend on the wavelet analysis code included in the MassSpecWavelet package [3]. All code herein was obtained under the GPL2 license and remains under the GPL 3 license or greater.
}
\seealso{
\link{\code{wave}} \link[dest=https://en.wikipedia.org/wiki/Continuous_wavelet_transform]{Continuous Wavelet Transform}

[1] Tautenhahn, R., Böttcher, C., & Neumann, S. (2008). Highly sensitive feature detection for high resolution LC/MS. BMC bioinformatics, 9(1), 1. \link[dest=http://dx.doi.org/10.1186/1471-2105-9-504]{http://dx.doi.org/10.1186/1471-2105-9-504}
[2] Smith, C. A., Want, E. J., O'Maille, G., Abagyan, R., & Siuzdak, G. (2006). XCMS: processing mass spectrometry data for metabolite profiling using nonlinear peak alignment, matching, and identification. Analytical chemistry, 78(3), 779-787. \link[dest=10.1021/ac051437y]{10.1021/ac051437y}
[3] Du, P., Kibbe, W. A., & Lin, S. M. (2006). Improved peak detection in mass spectrum by incorporating continuous wavelet transform-based pattern matching. Bioinformatics, 22(17), 2059-2065. \link[dest=10.1093/bioinformatics/btl355]{10.1093/bioinformatics/btl355}
}

