% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_baseline_noise.R
\name{estimateBaselineNoise}
\alias{estimateBaselineNoise}
\title{From an EIC estimate the local noise level and baseline of the signal}
\usage{
estimateBaselineNoise(eic, peakwidth)
}
\arguments{
\item{eic}{data.frame A data frame with columns "i" and "rt"}

\item{peakwidth}{integer A two number vector containing the min and max expected peakwidth.  Used to determine the smoothing window.}

\item{minslope.peak}{numeric The maximum slope a region may have to be considered a baseline region}
}
\value{
The original data.frame with additional columns i.sg - the smoothed EIC, baseline.region - 0 for peak regions, 1 for baseline regions, baseline - the estimated baseline intensity at that point, noise.sd - the estimated noise level at that point.
}
\description{
\code{estimateBaselineNoise} returns an EIC with additional vectors containing the estimated baseline at each point, the estimated noise level, and a smoothed EIC.
}
\details{
Baseline estimation is performed by excluding all smoothed regions with a slope less than \code{minslope.peak} and interpolating the remaining "baseline" regions.

Local noise estimation is performed by taking the standard deviation of the difference between the smoothed and the unsmoothed EICs.
}
\seealso{
\code{\link{wave}}
}

