% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_noise.R
\name{findLorentzianNoise}
\alias{findLorentzianNoise}
\title{Detect lorentzian noise around tall peaks in Orbitrap data
\code{findLorentzianNoise} returns the indices of peaks which fall below the noise threshold}
\usage{
findLorentzianNoise(p, lorentzian, peaks)
}
\arguments{
\item{p}{Named vector containing peak details}

\item{lorentzian}{function accepting a mass and returning the intensity of the noise at that mass}

\item{peaks}{The peak table to check for noise.}
}
\value{
Indices of peaks which are below the lorentzian noise threshold
}
\description{
Detect lorentzian noise around tall peaks in Orbitrap data
\code{findLorentzianNoise} returns the indices of peaks which fall below the noise threshold
}

