% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_noise.R
\name{findNoise}
\alias{findNoise}
\title{Takes a peak table and filters searches for black bar artifacts and lorentzian noise
\code{findNoise} returns the indices of peaks which are identified as noise}
\usage{
findNoise(peaks, int, resolution.calc, profxr, mirror.ppm = 2,
  mirror.logintrat = 0.7, mirror.totintrat = 0.1)
}
\arguments{
\item{peaks}{The peak table to filter for noise}

\item{int}{The intensity at which to check around a peak for noise}

\item{resolution.calc}{a function which accepts mass and returns the resolution}

\item{profxr}{an xcmsRaw object containing the profile mode data from this experiment.}

\item{mirror.ppm}{the ppm window in which to detect mirror artifacts}

\item{mirror.loginrat}{the intensity window in which to detect mirror artifacts}

\item{mirror.totinrat}{the maximum intensity relative to the tall peak which can be considered noise}
}
\value{
The indices in \code{peaks} which qualify as noise
}
\description{
Takes a peak table and filters searches for black bar artifacts and lorentzian noise
\code{findNoise} returns the indices of peaks which are identified as noise
}

