% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_noise.R
\name{findShoulderPeaks}
\alias{findShoulderPeaks}
\title{Detect artifactual shoulder peaks in Orbitrap data}
\usage{
findShoulderPeaks(profxr, p, ppm, logintrat = 0.7, totintrat = 0.01)
}
\arguments{
\item{profxr}{xcmsRaw object of the profile data}

\item{p}{named vector containng peak details such as mass, and retention time}

\item{ppm}{Allowance for mirroring in the mz domain}

\item{logintrat}{Allowance for mirroring in the intensity domain}

\item{totintrat}{Maximum height relative to main peak of shoulder peaks}
}
\value{
A vector of masses
}
\description{
\code{findShoulderPeaks} returns a list of masses which are shoulder peaks of a supplied mass
}

