% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_noise.R
\name{getLorentzianPeakshape}
\alias{getLorentzianPeakshape}
\title{Takes a mass and resolutionand returns a lorenzian describing the noise level in that region
\code{getLorentzianPeakshape} returns the indices of peaks which fall below the noise threshold}
\usage{
getLorentzianPeakshape(mass, resolution)
}
\arguments{
\item{mass}{numeric The location of the lorentzian peak}

\item{resolution}{Actual (mass specific) resolution at that mass}
}
\value{
A function accepting mass and returning normalized intensity
}
\description{
Takes a mass and resolutionand returns a lorenzian describing the noise level in that region
\code{getLorentzianPeakshape} returns the indices of peaks which fall below the noise threshold
}

