% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_wave_peak.R
\name{plotWavePeak}
\alias{plotWavePeak}
\title{Plot a detected peak with various additional metadata}
\usage{
plotWavePeak(i, eic, peaks)
}
\arguments{
\item{i}{The index of the peak to be plotted corresponding to a row in \code{peaks}}

\item{eic}{The eic which the peak was detected in. The input to \code{wave}}

\item{peaks}{The output of \code{wave}}
}
\value{
A ggplot2 plot grob.
}
\description{
\code{plotWavePeak} plots the peak, the snoothed peak, the local noise level, baseline, as well as wavelet and descent bounds
}
\seealso{
\link{\code{wave}} \link{\code{estimateBaselineNoise}}
}

