% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wave_centWave.R
\name{wave}
\alias{wave}
\title{Detect peaks in intensity vs time space by continuous wavelet transform}
\usage{
wave(eic, peakwidth = c(20, 50), valleywidth.min = 10, sensitivity = 1,
  smooth = T)
}
\arguments{
\item{eic}{data.frame containing columns: i - the EIC intensities, i.sg - the smoothed EIC intensitites, rt - the retention time of the observation in seconds, inroi - 1 if the scan is part of the ROI being interrogated, baseline - the estimated baseline at each point, noise.sd - the estimated SD of the noise at each point.}

\item{peakwidth}{integer A two number vector containing the min and max expected peakwidth in seconds. \emph{Note, do not make the minimum peakwidth too small - a 20 second peakwidth can still detect 2 second peaks.  Smaller peakwidths limit the ability of the algorithm to distinguish noise from peakshape.  The only downside of larger peakwidths is discerning poorly resolved peaks.}}

\item{valleywidth.min}{integer The minimum width of a valley in seconds before the peak is split.  Stochastic fluctuations can generate multiple scan valleys which are challenging to discern.  It is suggested that this remain reasonably large unless there is very little noise.}

\item{sensitivity}{numeric Larger sensitivities will search for peaks in noisier regions as defined by noise.sd of the eic. This comes at the cost of additional processing time.}

\item{smooth}{boolean Substitutes \code{i} for \code{i.sg} for relevant calculations.}
}
\value{
A matrix, rows corresponding to peaks. 

This matrix has many columns.  Columns prefixed with "wavelet." are analyses performed on the wavelet suggested peak bounds.  Columns prefixed with "descent." are analyses performed on the integration bounds (found by descending to the lowest point which can result in very long tails.) Peakshape comparisons like warpgroup may benefit from using the wavelet bounds as they will contain the bulk of the mass of the peak.
Useful columns include wavelet.fold.above.waveletbaseline and descent.fold.above.descentbaseline which have shown good stratification of peak quality and can be used to filter noise.
}
\description{
\code{wave} returns a list of all peak regions detected in an EIC as well as characteristics such as signal to noise estimations and integration regions.
}
\details{
This function is a modification of the original centWave algorithm [1]. The modifications are extensive - this code has not been tested to the extent which centWave has.

Modifications include:
\enumerate {
\item The algorithm accepts EICs rather than xcmsRaw objects for flexibility.
\item The peak detection algorithm has been separated from noise estimation code. (Noise estimation can be performed prior to analysis)
\item Valley tracking to limit the aggregation of closely eluting peaks
\item The ability to substitute smoothed EICs in relevant regions of the analysis
\item New peak quality measurements. (In our hands these provide more informative peak filters than the original code's signal to noise estimation.)
}
}
\section{Attribution}{

This code was modified from the originally published centWave algorithm [1].  The code was orginally distributed and obtained under the GPL2 license via the xcms software package [2]. The original algorithms depend on the wavelet analysis code included in the MassSpecWavelet package [3]. All code herein was obtained under the GPL2 license and remains under the GPL 3 license or greater.
}
\seealso{
\link{\code{wave}} \link{\code{estimateBaselineNoise}}

[1] Tautenhahn, R., B?ttcher, C., & Neumann, S. (2008). Highly sensitive feature detection for high resolution LC/MS. BMC bioinformatics, 9(1), 1. \link[dest=http://dx.doi.org/10.1186/1471-2105-9-504]{http://dx.doi.org/10.1186/1471-2105-9-504}
[2] Smith, C. A., Want, E. J., O'Maille, G., Abagyan, R., & Siuzdak, G. (2006). XCMS: processing mass spectrometry data for metabolite profiling using nonlinear peak alignment, matching, and identification. Analytical chemistry, 78(3), 779-787. \link[dest=10.1021/ac051437y]{10.1021/ac051437y}
[3] Du, P., Kibbe, W. A., & Lin, S. M. (2006). Improved peak detection in mass spectrum by incorporating continuous wavelet transform-based pattern matching. Bioinformatics, 22(17), 2059-2065. \link[dest=10.1093/bioinformatics/btl355]{10.1093/bioinformatics/btl355}
}

